Add arrow with json structured types support
