+++ src/main/java/net/snowflake/client/core/SFJsonResultSet.java
@@ -4,12 +4,8 @@
 
 package net.snowflake.client.core;
 
-import static net.snowflake.client.jdbc.SnowflakeUtil.getJsonNodeStringValue;
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import java.math.BigDecimal;
 import java.sql.Array;
 import java.sql.Date;
@@ -17,27 +13,16 @@ import java.sql.SQLInput;
 import java.sql.Time;
 import java.sql.Timestamp;
 import java.sql.Types;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Spliterator;
-import java.util.Spliterators;
 import java.util.TimeZone;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 import net.snowflake.client.core.json.Converters;
 import net.snowflake.client.core.structs.StructureTypeHelper;
 import net.snowflake.client.jdbc.ErrorCode;
-import net.snowflake.client.jdbc.FieldMetadata;
-import net.snowflake.client.jdbc.SnowflakeColumnMetadata;
 import net.snowflake.client.log.SFLogger;
 import net.snowflake.client.log.SFLoggerFactory;
-import net.snowflake.client.util.Converter;
 
 /** Abstract class used to represent snowflake result set in json format */
 public abstract class SFJsonResultSet extends SFBaseResultSet {
   private static final SFLogger logger = SFLoggerFactory.getLogger(SFJsonResultSet.class);
-  private static final ObjectMapper OBJECT_MAPPER = ObjectMapperFactory.getObjectMapper();
 
   protected final Converters converters;
 
@@ -133,7 +118,7 @@ public abstract class SFJsonResultSet extends SFBaseResultSet {
   @Override
   public Array getArray(int columnIndex) throws SFException {
     Object obj = getObjectInternal(columnIndex);
-    return getArrayInternal((String) obj, columnIndex);
+    return getJsonArrayInternal((String) obj, columnIndex);
   }
 
   @Override
@@ -266,41 +251,28 @@ public abstract class SFJsonResultSet extends SFBaseResultSet {
 
   @Override
   @SnowflakeJdbcInternalApi
-  public SQLInput createSqlInputForColumn(Object input, int columnIndex, SFBaseSession session) {
-    JsonNode inputNode;
-    if (input instanceof JsonNode) {
-      inputNode = (JsonNode) input;
-    } else {
-      inputNode = OBJECT_MAPPER.convertValue(input, JsonNode.class);
-    }
-    return new JsonSqlInput(
-        inputNode,
-        session,
-        converters,
-        resultSetMetaData.getColumnMetadata().get(columnIndex - 1).getFields(),
-        sessionTimezone);
+  public SQLInput createSqlInputForColumn(
+      Object input, Class<?> parentObjectClass, int columnIndex, SFBaseSession session) {
+    return createJsonSqlInputForColumn(input, columnIndex, session);
   }
 
   @Override
   @SnowflakeJdbcInternalApi
   public Date convertToDate(Object object, TimeZone tz) throws SFException {
-    return (Date) converters.dateStringConverter(session).convert(object);
+    return convertStringToDate(object, tz);
   }
 
   @Override
   @SnowflakeJdbcInternalApi
   public Time convertToTime(Object object, int scale) throws SFException {
-    return (Time) converters.timeFromStringConverter(session).convert(object);
+    return convertStringToTime(object, scale);
   }
 
   @Override
   @SnowflakeJdbcInternalApi
   public Timestamp convertToTimestamp(
       Object object, int columnType, int columnSubType, TimeZone tz, int scale) throws SFException {
-    return (Timestamp)
-        converters
-            .timestampFromStringConverter(columnSubType, columnType, scale, session, null, tz)
-            .convert(object);
+    return convertStringToTimestamp(object, columnType, columnSubType, tz, scale);
   }
 
   private Timestamp getTimestamp(int columnIndex) throws SFException {
@@ -325,146 +297,5 @@ public abstract class SFJsonResultSet extends SFBaseResultSet {
     } catch (JsonProcessingException e) {
       throw new SFException(e, ErrorCode.INVALID_STRUCT_DATA);
     }
-  }
-
-  private SfSqlArray getArrayInternal(String obj, int columnIndex) throws SFException {
-    try {
-      SnowflakeColumnMetadata arrayMetadata =
-          resultSetMetaData.getColumnMetadata().get(columnIndex - 1);
-      FieldMetadata fieldMetadata = arrayMetadata.getFields().get(0);
-
-      int columnSubType = fieldMetadata.getType();
-      int columnType = ColumnTypeHelper.getColumnType(columnSubType, session);
-      int scale = fieldMetadata.getScale();
-
-      ArrayNode arrayNode = (ArrayNode) OBJECT_MAPPER.readTree(obj);
-      Iterator<JsonNode> nodeElements = arrayNode.elements();
-
-      switch (columnSubType) {
-        case Types.INTEGER:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.integerConverter(columnType))
-                  .toArray(Integer[]::new));
-        case Types.SMALLINT:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.smallIntConverter(columnType))
-                  .toArray(Short[]::new));
-        case Types.TINYINT:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.tinyIntConverter(columnType))
-                  .toArray(Byte[]::new));
-        case Types.BIGINT:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.bigIntConverter(columnType)).toArray(Long[]::new));
-        case Types.DECIMAL:
-        case Types.NUMERIC:
-          return new SfSqlArray(
-              columnSubType,
-              convertToFixedArray(
-                  getStream(nodeElements, converters.bigDecimalConverter(columnType))));
-        case Types.CHAR:
-        case Types.VARCHAR:
-        case Types.LONGNVARCHAR:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.varcharConverter(columnType, columnSubType, scale))
-                  .toArray(String[]::new));
-        case Types.BINARY:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.bytesConverter(columnType, scale))
-                  .toArray(Byte[][]::new));
-        case Types.FLOAT:
-        case Types.REAL:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.floatConverter(columnType)).toArray(Float[]::new));
-        case Types.DOUBLE:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.doubleConverter(columnType))
-                  .toArray(Double[]::new));
-        case Types.DATE:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.dateStringConverter(session))
-                  .toArray(Date[]::new));
-        case Types.TIME:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.timeFromStringConverter(session))
-                  .toArray(Time[]::new));
-        case Types.TIMESTAMP:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(
-                      nodeElements,
-                      converters.timestampFromStringConverter(
-                          columnSubType, columnType, scale, session, null, sessionTimezone))
-                  .toArray(Timestamp[]::new));
-        case Types.BOOLEAN:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.booleanConverter(columnType))
-                  .toArray(Boolean[]::new));
-        case Types.STRUCT:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.structConverter(OBJECT_MAPPER))
-                  .toArray(Map[]::new));
-        case Types.ARRAY:
-          return new SfSqlArray(
-              columnSubType,
-              getStream(nodeElements, converters.arrayConverter(OBJECT_MAPPER))
-                  .toArray(Map[][]::new));
-        default:
-          throw new SFException(
-              ErrorCode.FEATURE_UNSUPPORTED,
-              "Can't construct array for data type: " + columnSubType);
-      }
-    } catch (JsonProcessingException e) {
-      throw new SFException(e, ErrorCode.INVALID_STRUCT_DATA);
-    }
-  }
-
-  private Object[] convertToFixedArray(Stream inputStream) {
-    AtomicInteger bigDecimalCount = new AtomicInteger();
-    Object[] elements =
-        inputStream
-            .peek(
-                elem -> {
-                  if (elem instanceof BigDecimal) {
-                    bigDecimalCount.incrementAndGet();
-                  }
-                })
-            .toArray(
-                size -> {
-                  boolean shouldReturnAsBigDecimal = bigDecimalCount.get() > 0;
-                  Class<?> returnedClass = shouldReturnAsBigDecimal ? BigDecimal.class : Long.class;
-                  return java.lang.reflect.Array.newInstance(returnedClass, size);
-                });
-    return elements;
-  }
-
-  private Stream getStream(Iterator nodeElements, Converter converter) {
-    return StreamSupport.stream(
-            Spliterators.spliteratorUnknownSize(nodeElements, Spliterator.ORDERED), false)
-        .map(
-            elem -> {
-              try {
-                return convert(converter, (JsonNode) elem);
-              } catch (SFException e) {
-                throw new RuntimeException(e);
-              }
-            });
-  }
-
-  private static Object convert(Converter converter, JsonNode node) throws SFException {
-    String nodeValue = getJsonNodeStringValue(node);
-    return converter.convert(nodeValue);
   }
 }
