+++ src/main/java/net/snowflake/client/core/structs/StructureTypeHelper.java
@@ -5,8 +5,7 @@ import net.snowflake.client.core.SnowflakeJdbcInternalApi;
 @SnowflakeJdbcInternalApi
 public class StructureTypeHelper {
   private static final String STRUCTURED_TYPE_ENABLED_PROPERTY_NAME = "STRUCTURED_TYPE_ENABLED";
-  private static boolean structuredTypeEnabled =
-      Boolean.valueOf(System.getProperty(STRUCTURED_TYPE_ENABLED_PROPERTY_NAME));
+  private static boolean structuredTypeEnabled = true;
 
   public static boolean isStructureTypeEnabled() {
     return structuredTypeEnabled;
