--- src/test/java/net/snowflake/client/jdbc/NativeArrowResultSetStructuredTypesLatestIT.java
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2012-2024 Snowflake Computing Inc. All right reserved.
- */
-package net.snowflake.client.jdbc;
-
-public class NativeArrowResultSetStructuredTypesLatestIT extends ResultSetStructuredTypesLatestIT {
-  public NativeArrowResultSetStructuredTypesLatestIT() {
-    super(ResultSetFormatType.NATIVE_ARROW);
-  }
-}
